refactor: split components to separate concerns
